feat: added separated file for topic id call
