+++ src/api/apiGetTopicsData.ts
@@ -17,11 +17,3 @@ export const apiGetTopicsData = async (): Promise<TopicsDataInterface[]> => {
 		Promise.reject(error);
 	});
 };
-
-export const apiGetTopicById = async (
-	topicId: any
-): Promise<TopicsDataInterface> => {
-	return apiGetTopicsData().then((response) =>
-		response.find((topic) => topic.id == topicId)
-	);
-};
